src/scrapers/static_scraper.py
"""Static scraper using BeautifulSoup and httpx."""
from typing import List, Dict, Any
import httpx
from bs4 import BeautifulSoup

from .base_scraper import BaseScraper


class StaticScraper(BaseScraper):
    """Scraper for static pages using BeautifulSoup."""
    
    def __init__(self, company_config: Dict[str, Any], scraping_config: Dict[str, Any], **kwargs):
        super().__init__(company_config, scraping_config, **kwargs)
        self.client: httpx.AsyncClient = None
    
    async def setup(self):
        """Initialize HTTP client."""
        self.log_progress("Setting up HTTP client")
        self.client = httpx.AsyncClient(
            timeout=30.0,
            follow_redirects=True,
            headers={
                'User-Agent': self.scraping_config.get('user_agent', 'Mozilla/5.0...'),
            }
        )
    
    async def teardown(self):
        """Close HTTP client."""
        self.log_progress("Closing HTTP client")
        if self.client:
            await self.client.aclose()
    
    async def scrape(self) -> List[Dict[str, Any]]:
        """
        Scrape job positions from static pages.
        
        Returns:
            List of job dictionaries
        """
        self.start_timer()
        jobs = []
        
        try:
            self.log_progress(f"Fetching {self.careers_url}")
            response = await self.client.get(self.careers_url)
            response.raise_for_status()
            self.increment_stat("requests_made")
            
            soup = BeautifulSoup(response.text, 'lxml')
            
            # TODO: Implement job extraction using selectors
            # TODO: Implement pagination handling
            
            self.log_progress(f"Scraping completed - Found {len(jobs)} jobs")
            
        except Exception as e:
            self.log_error(f"Scraping failed: {str(e)}", error=e)
        finally:
            self.stop_timer()
        
        return jobs

