requirements.txt
# Web Scraping
playwright==1.40.0
playwright-stealth==2.0.0
selenium==4.16.0
undetected-chromedriver==3.5.4
beautifulsoup4==4.12.2
lxml==4.9.3
httpx==0.25.2

# LLM Integration
openai==1.6.1
anthropic==0.8.1
langchain==0.1.0
langchain-openai==0.0.2
instructor==0.4.5

# Database & Storage
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
hiredis==2.3.2

# Task Queue & Orchestration
celery==5.3.4
dramatiq==1.15.0
apscheduler==3.10.4

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.1.4
python-dateutil==2.8.2

# API
fastapi==0.108.0
uvicorn[standard]==0.25.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
fake-useragent==1.4.0
aiohttp==3.9.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.9
mypy==1.8.0
ipython==8.18.1

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.2

